+++ 22215/k.kozlov/shell/makefile
@@ -1,2 +1,2 @@
 main:
-	gcc -Wall jobs.c shell.c promptline.c parseline.c -o main.out
+	gcc -Wall jobs.c shell_cmds.c shell.c promptline.c parseline.c -o main.out
