+++ 22215/k.kozlov/shell/parseline.c
@@ -8,6 +8,8 @@
 #include <fcntl.h>
 
 #include "shell.h"
+#include "shell_structs.h"
+#include "jobs.h"
 
 static char *blankskip(char *);
 
@@ -21,6 +23,7 @@ Job* parseline(char *line)
 	char *currentLinePtr;
 	char error = 0;
 	static char delim[] = " \t|&<>;\n";
+	// TODO: добавить корректный парсинг аргументов кавычках
 
 	/* initialize  */
 	currentLinePtr = line;
